refactor directory & add stream test
