+++ src/openworldlib/synthesis/visual_generation/yume/yume_1p5/__init__.py
@@ -1,0 +1,1 @@
+from .textimage2video import Yume1p5TI2V
